+++ changelog.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Both Match and MultiMatch queries may now specify a 'fuzziness', the default stays 'auto'
 - Using the `field()` on the search builder you may now define specifically which field(s) should be retrieved for the documents
 - New function score compound queries that can replace the default boolean compound query
+- A static `debug` method to help you with the last executed query
+- Laravel Scout's `take()` method can be used to set the max amount of results
+- Text analysis (analyzers, tokenizers, filters, etc.) is now possible through index settings.
+- Prepare (parts of) your data before letting Elasticsearch index it.
 
 ### Changed
 - Sorting now uses the default Scout `orderBy()` method